Upload PDF of waveform analysis
The resulting waveform was captured when plugging a leftover material spool eeprom chip into a material bay on a Stratasys F370 printer. This waveform was captured using a Tektronix oscilloscope set to 5M samples. The waveform captured was 10 s in width, which is more than enough time to capture the full communication between the eeprom and the printer. When plugging the eeprom into the material bay, the printer could almost immediately (within 2s) read the chip for the material type and % remaining on the spool from the eeprom. 

This proves that the waveform captured at this time WOULD contain the behavior/interaction of the eeprom and printer mainboard that allows this information to be read from the eeprom. If I can narrow down this behavior (which I believe I have) then I can then write a program to behave the same way as the printer mainboard and thus readout this information from the eeprom chip.  If I find exactly where this information (spool type and spool material remaining) is stored on the eeprom, then I can potentially write a program to edit it and reset an eeprom chip to 100%.